+++ eslint.config.js
@@ -31,12 +31,13 @@ export default tseslint.config(
     languageOptions: {
       globals: {
         process: "readonly",
-        console: "readonly"
+        console: "readonly",
+        Buffer: "readonly"
       }
     }
   },
   {
-    files: ["**/*.svelte"],
+    files: ["**/*.svelte", "**/*.svelte.ts", "**/*.svelte.js"],
     languageOptions: {
       parserOptions: {
         parser: tseslint.parser,